Create Basic principles
Specially trained judges for justice and the development of an artificial intelligence-supported system that monitors the correct application of the laws that form the basis of judges' decisions.
Families who want to obtain the rank of judge, the most respected position in the community, enroll their children in justice school by passing special interviews, exams and security exams.
To prevent these families from being manipulated by criminals, the family waives its right to personal protection.
Judges trained in this school rule with an advanced algorithm and system supported by artificial intelligence. The powers of any political office are not superior to the decisions of the judges. Everyone is equal before the law. Including MPs and ministers. Judges are judged only by facts.